Adhere to the new dictionary interface with `get`
The problem is introduced commit
https://github.com/openstenoproject/plover/commit/1fc1e313c724fe0d8676e35294296cb196f79c0b,
where it adds a `get` method that delegates to the dictionary. This
dictionary doesn't have a `get` method, so it uses the default
implementation, which is wrong here.

This could also be fixed by having `StenoDictionary.get` call
`StenoDictionary.__getitem__` by default, since we currently do the
right thing for `__getitem__` in this dictionary.